src/fashion-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
scikit-learn
numpy
pillow
transformers
chromadb
annoy
scipy